Refactor the partner field amdin model
