+++ src/django/api/admin.py
@@ -1,7 +1,6 @@
 import json
 import logging
 
-from django import forms
 from django.db import transaction
 from django.urls import path
 from django.contrib import admin, messages
@@ -14,7 +13,6 @@ from django.urls import reverse
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 from api.models.sector_group import SectorGroup
-from api.models.partner_field import PartnerField
 from api.models.partner_field_group import PartnerFieldGroup
 from api.models.wage_indicator_country_data import WageIndicatorCountryData
 from api.models.wage_indicator_link_text_config import (
@@ -30,9 +28,10 @@ from allauth.account.models import EmailAddress
 from simple_history.admin import SimpleHistoryAdmin
 from waffle.models import Flag, Sample, Switch
 from waffle.admin import FlagAdmin, SampleAdmin, SwitchAdmin
-from jsoneditor.forms import JSONEditor
 
 from api import models
+from api.models.partner_field import PartnerField
+from api.models.partner_field_admin import PartnerFieldAdmin
 
 from api.reports import get_report_names, run_report
 
@@ -273,137 +272,6 @@ class SectorGroupAdmin(admin.ModelAdmin):
 
     def get_ordering(self, request):
         return ['name']
-
-
-class PartnerFieldAdminForm(forms.ModelForm):
-    json_schema = forms.JSONField(
-        required=False,
-        widget=JSONEditor(
-            init_options={"mode": "code", "modes": ["code", "tree"]},
-            attrs={
-                'style': 'width: 800px; height: 400px;'
-            }
-        )
-    )
-
-    class Meta:
-        model = PartnerField
-        fields = [
-            'name',
-            'type',
-            'unit',
-            'label',
-            'source_by',
-            'base_url',
-            'display_text',
-            'json_schema',
-            'active',
-            'system_field'
-        ]
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def clean(self):
-        '''
-        Validate that protected fields of system fields are not modified.
-        '''
-        cleaned_data = super().clean()
-
-        if self.instance and self.instance.pk and self.instance.system_field:
-            try:
-                original = PartnerField.objects.get(pk=self.instance.pk)
-
-                protected_fields = {
-                    'name': 'Name',
-                    'type': 'Type',
-                    'json_schema': 'JSON Schema',
-                    'system_field': 'System Field'
-                }
-
-                for field_name, field_label in protected_fields.items():
-                    original_value = getattr(original, field_name)
-                    current_value = cleaned_data.get(field_name)
-
-                    if original_value != current_value:
-                        self.add_error(
-                            field_name,
-                            f'{field_label} cannot be modified for '
-                            'system-defined fields. Editing this field may '
-                            'break the application or data display for users. '
-                            'Only label, unit, source by, base url, display '
-                            'text, and active can be edited.'
-                        )
-
-            except PartnerField.DoesNotExist:
-                logger.warning(
-                    f'Partner field `{self.instance.pk}` not found. '
-                    'System field must exist in database.'
-                )
-
-        return cleaned_data
-
-
-class PartnerFieldAdmin(admin.ModelAdmin):
-    form = PartnerFieldAdminForm
-    list_display = (
-        "name",
-        "type",
-        "label",
-        "unit",
-        "group",
-        "active",
-        "system_field",
-        "created_at",
-    )
-    search_fields = ("name", "type", "label", "unit", "source_by")
-    list_filter = ("active", "system_field", "type", "group")
-    readonly_fields = ("uuid", "created_at", "updated_at")
-    fields = (
-        "name",
-        "type",
-        "unit",
-        "label",
-        "group",
-        "source_by",
-        "base_url",
-        "display_text",
-        "json_schema",
-        "active",
-        "system_field",
-        "created_at",
-        "updated_at",
-    )
-
-    def get_queryset(self, request):
-        '''
-        Override to show all partner fields including inactive ones in admin.
-        '''
-        qs = self.model.objects.get_all_including_inactive()
-        ordering = self.get_ordering(request)
-        if ordering:
-            qs = qs.order_by(*ordering)
-        return qs
-
-    def has_delete_permission(self, request, obj=None):
-        '''
-        Prevent deletion of system fields.
-        '''
-        if obj and obj.system_field:
-            messages.warning(
-                request,
-                f'Partner field \'{obj.name}\' cannot be deleted because it '
-                'is a system-defined field.'
-            )
-            return False
-
-        return super().has_delete_permission(request, obj)
-
-    class Media:
-        js = (
-            'admin/js/jquery.init.js',
-            'admin/js/partner_field_admin.js',
-        )
 
 
 class EmailAddressAdmin(admin.ModelAdmin):
